tests/wpt/tests/navigation-api/commit-behavior/after-transition-traversal-window-stop-before-commit.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../navigation-methods/return-value/resources/helpers.js"></script>
<body>
<script>
promise_test(async t => {
  // Wait for after the load event because window.stop() hangs the test harness
  // if called before the load event.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));

  let start_length = navigation.entries().length;
  let start_index = navigation.currentEntry.index;

  await navigation.navigate("#1").finished;
  await navigation.navigate("#2").finished;
  assert_equals(navigation.entries().length, start_length + 2);
  assert_equals(navigation.currentEntry.index, start_index + 2);

  let stopped_first_traverse = false;
  navigation.onnavigate = t.step_func(e => {
    if (!stopped_first_traverse) {
      stopped_first_traverse = true;
      t.step_timeout(() => window.stop(), 0);
    }
    e.intercept({ commit: "after-transition",
                  handler: () => new Promise(r => t.step_timeout(r, 10)) });
  });

  let navigatesuccess_called = false;
  let navigateerror_called = false;
  navigation.onnavigatesuccess = t.step_func(() => {
    navigatesuccess_called = true;
    assert_equals(location.hash, "#1");
  });
  navigation.onnavigateerror = t.step_func(() => {
    navigateerror_called = true;
    assert_equals(location.hash, "#2");
  });

  await assertBothRejectDOM(t, navigation.back(), "AbortError");

  assert_equals(navigation.currentEntry.index, start_index + 2);
  assert_equals(location.hash, "#2");
  assert_true(navigateerror_called);
  assert_false(navigatesuccess_called);

  // Ensure navigation.entries() remains in sync with the session history in
  // spite of the window.stop(), and navigation.back() does not skip over #1.
  await navigation.back().finished;

  assert_equals(navigation.currentEntry.index, start_index + 1);
  assert_equals(location.hash, "#1");
  assert_true(navigatesuccess_called);
}, " { commit: 'after-transition' } traverse with window.stop() before commit");
</script>
</body>
